+++ app/src/main/kotlin/com/simplemobiletools/calendar/pro/helpers/CalDAVHelper.kt
@@ -183,8 +183,8 @@ class CalDAVHelper(val context: Context) {
                     val id = cursor.getLongValue(CalendarContract.Events._ID)
                     val title = cursor.getStringValue(CalendarContract.Events.TITLE) ?: ""
                     val description = cursor.getStringValue(CalendarContract.Events.DESCRIPTION) ?: ""
-                    val startTS = (cursor.getLongValue(CalendarContract.Events.DTSTART) / 1000).toInt()
-                    var endTS = (cursor.getLongValue(CalendarContract.Events.DTEND) / 1000).toInt()
+                    val startTS = cursor.getLongValue(CalendarContract.Events.DTSTART) / 1000L
+                    var endTS = cursor.getLongValue(CalendarContract.Events.DTEND) / 1000L
                     val allDay = cursor.getIntValue(CalendarContract.Events.ALL_DAY)
                     val rrule = cursor.getStringValue(CalendarContract.Events.RRULE) ?: ""
                     val location = cursor.getStringValue(CalendarContract.Events.EVENT_LOCATION) ?: ""
@@ -192,7 +192,7 @@ class CalDAVHelper(val context: Context) {
                     val originalInstanceTime = cursor.getLongValue(CalendarContract.Events.ORIGINAL_INSTANCE_TIME)
                     val reminders = getCalDAVEventReminders(id)
 
-                    if (endTS == 0) {
+                    if (endTS == 0L) {
                         val duration = cursor.getStringValue(CalendarContract.Events.DURATION) ?: ""
                         endTS = startTS + Parser().parseDurationSeconds(duration)
                     }
@@ -234,7 +234,7 @@ class CalDAVHelper(val context: Context) {
                             val parentEventId = context.eventsDB.getEventIdWithImportId(parentImportId)
                             if (parentEventId != null) {
                                 event.parentId = parentEventId
-                                eventsHelper.addEventRepeatException(parentEventId, (originalInstanceTime / 1000).toInt(), false, event.importId)
+                                eventsHelper.addEventRepeatException(parentEventId, originalInstanceTime / 1000L, false, event.importId)
                             }
                         }
 
@@ -353,7 +353,7 @@ class CalDAVHelper(val context: Context) {
             val dur = Math.max(1, (event.endTS - event.startTS) / DAY)
             "P${dur}D"
         } else {
-            Parser().getDurationCode((event.endTS - event.startTS) / 60)
+            Parser().getDurationCode((event.endTS - event.startTS) / 60L)
         }
     }
 
@@ -373,7 +373,7 @@ class CalDAVHelper(val context: Context) {
     }
 
     @SuppressLint("MissingPermission")
-    fun insertEventRepeatException(event: Event, occurrenceTS: Int): Long {
+    fun insertEventRepeatException(event: Event, occurrenceTS: Long): Long {
         val uri = CalendarContract.Events.CONTENT_URI
         val values = fillEventRepeatExceptionValues(event, occurrenceTS)
         val newUri = context.contentResolver.insert(uri, values)
@@ -381,7 +381,7 @@ class CalDAVHelper(val context: Context) {
         return java.lang.Long.parseLong(newUri.lastPathSegment)
     }
 
-    private fun fillEventRepeatExceptionValues(event: Event, occurrenceTS: Int): ContentValues {
+    private fun fillEventRepeatExceptionValues(event: Event, occurrenceTS: Long): ContentValues {
         return ContentValues().apply {
             put(CalendarContract.Events.CALENDAR_ID, event.getCalDAVCalendarId())
             put(CalendarContract.Events.DTSTART, occurrenceTS)
