+++ app/src/main/kotlin/com/simplemobiletools/calendar/pro/models/Event.kt
@@ -15,8 +15,8 @@ import java.io.Serializable
 @Entity(tableName = "events", indices = [(Index(value = ["id"], unique = true))])
 data class Event(
         @PrimaryKey(autoGenerate = true) var id: Long?,
-        @ColumnInfo(name = "start_ts") var startTS: Int = 0,
-        @ColumnInfo(name = "end_ts") var endTS: Int = 0,
+        @ColumnInfo(name = "start_ts") var startTS: Long = 0L,
+        @ColumnInfo(name = "end_ts") var endTS: Long = 0L,
         @ColumnInfo(name = "title") var title: String = "",
         @ColumnInfo(name = "location") var location: String = "",
         @ColumnInfo(name = "description") var description: String = "",
@@ -25,7 +25,7 @@ data class Event(
         @ColumnInfo(name = "reminder_3_minutes") var reminder3Minutes: Int = -1,
         @ColumnInfo(name = "repeat_interval") var repeatInterval: Int = 0,
         @ColumnInfo(name = "repeat_rule") var repeatRule: Int = 0,
-        @ColumnInfo(name = "repeat_limit") var repeatLimit: Int = 0,
+        @ColumnInfo(name = "repeat_limit") var repeatLimit: Long = 0L,
         @ColumnInfo(name = "import_id") var importId: String = "",
         @ColumnInfo(name = "flags") var flags: Int = 0,
         @ColumnInfo(name = "event_type") var eventType: Long = REGULAR_EVENT_TYPE_ID,
@@ -118,7 +118,7 @@ data class Event(
     fun getReminders() = setOf(reminder1Minutes, reminder2Minutes, reminder3Minutes).filter { it != REMINDER_OFF }
 
     // properly return the start time of all-day events as midnight
-    fun getEventStartTS(): Int {
+    fun getEventStartTS(): Long {
         return if (getIsAllDay()) {
             Formatter.getDateTimeFromTS(startTS).withTime(0, 0, 0, 0).seconds()
         } else {
@@ -139,7 +139,7 @@ data class Event(
     fun getEventRepetition() = EventRepetition(null, id!!, repeatInterval, repeatRule, repeatLimit)
 
     // check if its the proper week, for events repeating every x weeks
-    fun isOnProperWeek(startTimes: LongSparseArray<Int>): Boolean {
+    fun isOnProperWeek(startTimes: LongSparseArray<Long>): Boolean {
         val initialWeekOfYear = Formatter.getDateTimeFromTS(startTimes[id!!]!!).weekOfWeekyear
         val currentWeekOfYear = Formatter.getDateTimeFromTS(startTS).weekOfWeekyear
         return (currentWeekOfYear - initialWeekOfYear) % (repeatInterval / WEEK) == 0
